pyproject.toml
[project]
name = "playlists"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "pillow>=11.3.0",
    "spotipy>=2.25.1",
]
